Allows buildpack output to stream commands
The `start_stream` interface only returns a single writer. The Command expects two writers: stdout and stderr. There's currently no good way for a buildpack author to stream a running Command, to the point that there are no examples of it in the documentation.

This PR introduces a mechanism to do this and adds a doc that executes a command streaming to the build output.