Add a way to cancel a timer
When dealing with network IO it's expected that we will need to retry. There's currently no good way to advertise what's happening to a user.

This new public API allows for canceling of running timers (without emitting "Done") after which time the timer can be restarted. That allows for an output that could look like this:


```
- JVM version 8.0.0
  - Downloading JVM .... (retrying in 1 second)
  - Retrying ....... (retrying in 2 seconds)
  - Retrying ...... (Done 4s)
```