Unneeded perf optimization
We know that the string that is produced will be at MOST the same size as the &str passed in. If we allocate that up-front then we avoid potentially needing to allocate several times while iterating over the chars.

It's not required in any way shape or form, but since we can do it and it's easy enough, why not!